+++ CHANGELOG.md
@@ -1,3 +1,109 @@
+## 2.0.0-alpha.1 (2021-01-25) REFACTOR
+
+## Chore
+
+- @capacitor/core@3.0.0-beta.1
+- @capacitor-community/sqlite@3.0.0-beta.2
+
+## 1.0.0 (2021-01-25) REFACTOR
+
+## Chore
+
+- Move to branch 2.9.x tag latest
+- @capacitor-community/sqlite@2.9.6
+
+## 1.0.0-beta.4 (2021-01-14) REFACTOR
+
+## Chore
+
+- @capacitor/core@2.4.6
+- @capacitor-community/sqlite@2.9.0
+
+### Bug Fixes
+
+- Fix issue#60 remove Android permissions
+- Fix readme & changelog
+- fix APIHook.md
+
+## 1.0.0-beta.3 (2021-01-05) REFACTOR
+
+## Chore
+
+- @capacitor-community/sqlite@2.9.0-beta.3
+
+### Bug Fixes
+
+- Fix readme & changelog
+- fix APIHook.md
+
+## 1.0.0-beta.2 (2021-01-01) REFACTOR
+
+### Bug Fixes
+
+- Fix readme
+- Fix test
+
+## 1.0.0-beta.1 (2020-12-28) REFACTOR
+
+## Chore
+
+- @capacitor/core@2.4.5
+- @capacitor-community/sqlite@2.9.0-beta.1
+
+## Added Features
+
+- Add importFromJson and isJsonValid methods
+
+## 1.0.0-alpha.4 (2020-12-05) REFACTOR
+
+### Bug Fixes
+
+- Fix readme
+
+## 1.0.0-alpha.3 (2020-12-04) REFACTOR
+
+### Added Features
+
+ - getPlatform
+
+### Bug Fixes
+
+ - fix new SQLiteConnection with useMemo()
+
+
+## 1.0.0-alpha.2 (2020-12-03) REFACTOR
+
+### Chore
+
+- @capacitor-community/sqlite@2.9.0-alpha.5
+
+## 1.0.0-alpha.1 (2020-12-02) REFACTOR
+
+### Chore
+
+- @capacitor-community/sqlite@refactor
+
+### Added Features
+
+ - echo
+ - createConnection
+ - closeConnection
+ - retrieveConnection
+ - closeAllConnections
+ - retrieveAllConnections
+ - addUpgradeStatement
+ - requestPermissions
+
+## 0.0.9 (2021-01-25)
+
+### Chores
+
+- Has been moved to tag 'INITIAL' not maintained anymore
+
+### Bug Fixes
+
+- Update README
+
 ## 0.0.8 (2021-01-13)
 
 ### Bug Fixes